+++ mandatory/srcs/Mlx/usefull.c
@@ -6,13 +6,13 @@
 /*   By: lde-merc <lde-merc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/13 16:02:27 by lde-merc          #+#    #+#             */
-/*   Updated: 2025/05/28 12:24:50 by lde-merc         ###   ########.fr       */
+/*   Updated: 2025/06/16 17:16:14 by lde-merc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
-float	ft_distance(t_position *pos1, t_position *pos2)
+double	ft_distance(t_position *pos1, t_position *pos2)
 {
 	return (sqrt(pow((pos2->x - pos1->x), 2) + pow((pos2->y - pos1->y), 2)));
 }
